+++ google-maps/src/definitions.ts
@@ -1,10 +1,32 @@
+import { CapacitorGoogleMaps } from './implementation';
+
 /**
  * An interface representing the viewports latitude and longitude bounds.
  */
-export interface LatLngBounds {
+export interface LatLngBoundsInterface {
   southwest: LatLng;
   center: LatLng;
   northeast: LatLng;
+}
+
+export class LatLngBounds {
+  southwest: LatLng;
+  center: LatLng;
+  northeast: LatLng;
+
+  constructor(bounds: LatLngBoundsInterface) {
+    this.southwest = bounds.southwest;
+    this.center = bounds.center;
+    this.northeast = bounds.northeast;
+  }
+
+  async contains(point: LatLng): Promise<boolean> {
+    const result = await CapacitorGoogleMaps.mapBoundsContains({
+      bounds: this,
+      point,
+    });
+    return result['contains'];
+  }
 }
 
 /**
@@ -33,9 +55,11 @@ export interface Point {
 }
 
 /**
- *
+ * For web, all the javascript Google Maps options are available as
+ * GoogleMapConfig extends google.maps.MapOptions.
+ * For iOS and Android only the config options declared on GoogleMapConfig are available.
  */
-export interface GoogleMapConfig {
+export interface GoogleMapConfig extends google.maps.MapOptions {
   /**
    * Override width for native map.
    */
@@ -70,6 +94,14 @@ export interface GoogleMapConfig {
    * Override pixel ratio for native map.
    */
   devicePixelRatio?: number;
+  /**
+   * Styles to apply to each of the default map types. Note that for
+   * satellite, hybrid and terrain modes,
+   * these styles will only apply to labels and geometry.
+   *
+   * @since 4.3.0
+   */
+  styles?: google.maps.MapTypeStyle[] | null;
 }
 
 /**
@@ -106,7 +138,7 @@ export interface CameraConfig {
   animate?: boolean;
 
   /**
-   *
+   * This configuration option is not being used.
    */
   animationDuration?: number;
 }
@@ -230,6 +262,14 @@ export interface Marker {
    * @default false
    */
   draggable?: boolean;
+
+  /**
+   * Specifies the stack order of this marker, relative to other markers on the map.
+   * A marker with a high z-index is drawn on top of markers with lower z-indexes
+   *
+   * @default 0
+   */
+  zIndex?: number;
 }
 
 /**