+++ local-notifications/android/src/main/java/com/capacitorjs/plugins/localnotifications/TimedNotificationPublisher.java
@@ -28,7 +28,15 @@ public class TimedNotificationPublisher extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
         NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-        Notification notification = intent.getParcelableExtra(NOTIFICATION_KEY);
+
+        Notification notification;
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            notification = intent.getParcelableExtra(NOTIFICATION_KEY, Notification.class);
+        } else {
+            notification = getParcelableExtraLegacy(intent, NOTIFICATION_KEY);
+        }
+
         notification.when = System.currentTimeMillis();
 
         int id = intent.getIntExtra(LocalNotificationManager.NOTIFICATION_INTENT_KEY, Integer.MIN_VALUE);
@@ -42,6 +50,11 @@ public class TimedNotificationPublisher extends BroadcastReceiver {
         if (!rescheduleNotificationIfNeeded(context, intent, id)) {
             storage.deleteNotification(Integer.toString(id));
         }
+    }
+
+    @SuppressWarnings("deprecation")
+    private Notification getParcelableExtraLegacy(Intent intent, String string) {
+        return intent.getParcelableExtra(NOTIFICATION_KEY);
     }
 
     private boolean rescheduleNotificationIfNeeded(Context context, Intent intent, int id) {