+++ push-notifications/src/definitions.ts
@@ -93,7 +93,7 @@ export interface PushNotificationsPlugin {
   /**
    * Check permission to receive push notifications.
    *
-   * On Android the status is always granted because you can always
+   * On Android 12 and below the status is always granted because you can always
    * receive push notifications. If you need to check if the user allows
    * to display notifications, use local-notifications plugin.
    *
@@ -104,12 +104,12 @@ export interface PushNotificationsPlugin {
   /**
    * Request permission to receive push notifications.
    *
-   * On Android it doesn't prompt for permission because you can always
+   * On Android 12 and below it doesn't prompt for permission because you can always
    * receive push notifications.
    *
    * On iOS, the first time you use the function, it will prompt the user
    * for push notification permission and return granted or denied based
-   * on the user selection. On following calls it will currect status of
+   * on the user selection. On following calls it will get the current status of
    * the permission without prompting again.
    *
    * @since 1.0.0