+++ splash-screen/android/src/main/java/com/capacitorjs/plugins/splashscreen/SplashScreen.java
@@ -12,6 +12,7 @@ import android.graphics.PixelFormat;
 import android.graphics.drawable.Animatable;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.LayerDrawable;
+import android.os.Build;
 import android.os.Handler;
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -19,6 +20,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewTreeObserver;
 import android.view.ViewTreeObserver.OnPreDrawListener;
+import android.view.Window;
+import android.view.WindowInsetsController;
 import android.view.WindowManager;
 import android.view.animation.LinearInterpolator;
 import android.widget.FrameLayout;
@@ -26,6 +29,8 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.ProgressBar;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.view.WindowCompat;
+import androidx.core.view.WindowInsetsCompat;
 import com.getcapacitor.Logger;
 
 /**
@@ -97,29 +102,44 @@ public class SplashScreen {
                 );
                 windowSplashScreen.setKeepOnScreenCondition(() -> isVisible || isHiding);
 
-                // Set Fade Out Animation
-                windowSplashScreen.setOnExitAnimationListener(
-                    windowSplashScreenView -> {
-                        final ObjectAnimator fadeAnimator = ObjectAnimator.ofFloat(windowSplashScreenView.getView(), View.ALPHA, 1f, 0f);
-                        fadeAnimator.setInterpolator(new LinearInterpolator());
-                        fadeAnimator.setDuration(settings.getFadeOutDuration());
+                if (config.getLaunchFadeOutDuration() > 0) {
+                    // Set Fade Out Animation
+                    windowSplashScreen.setOnExitAnimationListener(
+                        windowSplashScreenView -> {
+                            final ObjectAnimator fadeAnimator = ObjectAnimator.ofFloat(
+                                windowSplashScreenView.getView(),
+                                View.ALPHA,
+                                1f,
+                                0f
+                            );
+                            fadeAnimator.setInterpolator(new LinearInterpolator());
+                            fadeAnimator.setDuration(config.getLaunchFadeOutDuration());
 
-                        fadeAnimator.addListener(
-                            new AnimatorListenerAdapter() {
-                                @Override
-                                public void onAnimationEnd(Animator animation) {
-                                    isHiding = false;
-                                    windowSplashScreenView.remove();
+                            fadeAnimator.addListener(
+                                new AnimatorListenerAdapter() {
+                                    @Override
+                                    public void onAnimationEnd(Animator animation) {
+                                        isHiding = false;
+                                        windowSplashScreenView.remove();
+                                    }
                                 }
-                            }
-                        );
+                            );
 
-                        fadeAnimator.start();
+                            fadeAnimator.start();
 
-                        isHiding = true;
-                        isVisible = false;
-                    }
-                );
+                            isHiding = true;
+                            isVisible = false;
+                        }
+                    );
+                } else {
+                    windowSplashScreen.setOnExitAnimationListener(
+                        windowSplashScreenView -> {
+                            isHiding = false;
+                            isVisible = false;
+                            windowSplashScreenView.remove();
+                        }
+                    );
+                }
 
                 // Set Pre Draw Listener & Delay Drawing Until Duration Elapses
                 content = activity.findViewById(android.R.id.content);
@@ -222,7 +242,7 @@ public class SplashScreen {
                 isVisible = true;
 
                 if (settings.isAutoHide()) {
-                    new Handler()
+                    new Handler(context.getMainLooper())
                         .postDelayed(
                             () -> {
                                 hideDialog(activity, isLaunchSplash);
@@ -312,26 +332,17 @@ public class SplashScreen {
                     // Stops flickers dead in their tracks
                     // https://stackoverflow.com/a/21847579/32140
                     ImageView imageView = (ImageView) splashImage;
-                    imageView.setDrawingCacheEnabled(true);
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                        imageView.setLayerType(View.LAYER_TYPE_SOFTWARE, null);
+                    } else {
+                        legacyStopFlickers(imageView);
+                    }
                     imageView.setScaleType(config.getScaleType());
                     imageView.setImageDrawable(splash);
                 }
             }
 
             splashImage.setFitsSystemWindows(true);
-
-            if (config.isImmersive()) {
-                final int flags =
-                    View.SYSTEM_UI_FLAG_LAYOUT_STABLE |
-                    View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION |
-                    View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN |
-                    View.SYSTEM_UI_FLAG_HIDE_NAVIGATION |
-                    View.SYSTEM_UI_FLAG_FULLSCREEN |
-                    View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY;
-                splashImage.setSystemUiVisibility(flags);
-            } else if (config.isFullScreen()) {
-                splashImage.setSystemUiVisibility(View.SYSTEM_UI_FLAG_FULLSCREEN);
-            }
 
             if (config.getBackgroundColor() != null) {
                 splashImage.setBackgroundColor(config.getBackgroundColor());
@@ -360,6 +371,11 @@ public class SplashScreen {
                 spinnerBar.setIndeterminateTintList(colorStateList);
             }
         }
+    }
+
+    @SuppressWarnings("deprecation")
+    private void legacyStopFlickers(ImageView imageView) {
+        imageView.setDrawingCacheEnabled(true);
     }
 
     private Drawable getSplashDrawable() {
@@ -398,7 +414,7 @@ public class SplashScreen {
                 isVisible = true;
 
                 if (settings.isAutoHide()) {
-                    new Handler()
+                    new Handler(context.getMainLooper())
                         .postDelayed(
                             () -> {
                                 hide(settings.getFadeOutDuration(), isLaunchSplash);
@@ -445,6 +461,35 @@ public class SplashScreen {
                     return;
                 }
 
+                if (config.isImmersive()) {
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+                        activity.runOnUiThread(
+                            () -> {
+                                Window window = activity.getWindow();
+                                WindowCompat.setDecorFitsSystemWindows(window, false);
+                                WindowInsetsController controller = splashImage.getWindowInsetsController();
+                                controller.hide(WindowInsetsCompat.Type.systemBars());
+                                controller.setSystemBarsBehavior(WindowInsetsController.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE);
+                            }
+                        );
+                    } else {
+                        legacyImmersive();
+                    }
+                } else if (config.isFullScreen()) {
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+                        activity.runOnUiThread(
+                            () -> {
+                                Window window = activity.getWindow();
+                                WindowCompat.setDecorFitsSystemWindows(window, false);
+                                WindowInsetsController controller = splashImage.getWindowInsetsController();
+                                controller.hide(WindowInsetsCompat.Type.statusBars());
+                            }
+                        );
+                    } else {
+                        legacyFullscreen();
+                    }
+                }
+
                 splashImage.setAlpha(0f);
 
                 splashImage
@@ -486,6 +531,23 @@ public class SplashScreen {
         );
     }
 
+    @SuppressWarnings("deprecation")
+    private void legacyImmersive() {
+        final int flags =
+            View.SYSTEM_UI_FLAG_LAYOUT_STABLE |
+            View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION |
+            View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN |
+            View.SYSTEM_UI_FLAG_HIDE_NAVIGATION |
+            View.SYSTEM_UI_FLAG_FULLSCREEN |
+            View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY;
+        splashImage.setSystemUiVisibility(flags);
+    }
+
+    @SuppressWarnings("deprecation")
+    private void legacyFullscreen() {
+        splashImage.setSystemUiVisibility(View.SYSTEM_UI_FLAG_FULLSCREEN);
+    }
+
     private void hide(final int fadeOutDuration, boolean isLaunchSplash) {
         // Warn the user if the splash was hidden automatically, which means they could be experiencing an app
         // that feels slower than it actually is.
@@ -503,6 +565,11 @@ public class SplashScreen {
 
         // Hide with Android 12 API
         if (null != this.onPreDrawListener) {
+            if (fadeOutDuration != 200) {
+                Logger.warn(
+                    "fadeOutDuration parameter doesn't work on initial splash screen, use launchFadeOutDuration configuration option"
+                );
+            }
             this.isVisible = false;
             if (null != content) {
                 content.getViewTreeObserver().removeOnPreDrawListener(this.onPreDrawListener);
@@ -614,6 +681,11 @@ public class SplashScreen {
             windowManager.removeView(splashImage);
         }
 
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R && config.isFullScreen() || config.isImmersive()) {
+            // Exit fullscreen mode
+            Window window = ((Activity) context).getWindow();
+            WindowCompat.setDecorFitsSystemWindows(window, true);
+        }
         isHiding = false;
         isVisible = false;
     }