+++ motion/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0-alpha.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/motion@4.0.2...@capacitor/motion@5.0.0-alpha.1) (2023-03-16)
+
+**Note:** Version bump only for package @capacitor/motion
+
+
+
+
+
+## [4.0.2](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/motion@1.0.5...@capacitor/motion@4.0.2) (2022-11-16)
+
+
+
+## 4.0.1 (2022-07-28)
+
+
+
+# 4.0.0 (2022-07-27)
+
+
+
+# 4.0.0-beta.2 (2022-07-08)
+
+
+
+# 4.0.0-beta.0 (2022-06-27)
+
+**Note:** Version bump only for package @capacitor/motion
+
+
+
+
+
 ## [4.0.1](https://github.com/ionic-team/capacitor-plugins/compare/4.0.0...4.0.1) (2022-07-28)
 
 **Note:** Version bump only for package @capacitor/motion